Cast Chesapeake data to `FloatTensor`

+++ datasets.py
@@ -31,8 +31,8 @@ class Chesapeake_CIFAR10(VisionDataset):
         image_path = os.path.join(sub_dir_path, self.image_fn)
         target_path = os.path.join(sub_dir_path, self.targets_fn)
         
-        self.data = torch.load(image_path)
-        self.targets = torch.load(target_path)
+        self.data = torch.load(image_path).type(torch.FloatTensor)
+        self.targets = torch.load(target_path).type(torch.FloatTensor)
         
         self.classes = [0, 1, 2, 3, 4, 5, 6, 7, 8]
